[Cascade-skip resurrection](2) Dispatch a resurrected descendant's launch outbox entry
Once (1) un-skips a descendant, orchestrator.approve()'s own
autoStartReadyTasks call enqueues it (status queued, phase launching)
but nothing in headless mode drains that launch-outbox entry: the
mutation-scoped/global-topup dispatch paths are both intentionally
no-ops ("the durable launch outbox owns dispatch"), and headless
approve's own "did new work start" heuristic only looked at tasks
still sitting in pending, so a task that moved skipped -> pending ->
queued within the same approve() call was invisible to it. The task
sat enqueued forever with no owner-serve process to pick it up.

dispatchHeadlessRunnableTasks (moved verbatim from headless.ts, no
behavior change) is the same local poller already used by the
edit-task-* commands; headlessApprove now runs it against the
mutation's topup tasks and folds non-running topup entries into its
resumedWork decision so it waits for them to settle.
